\name{SetUpProject-package}
\alias{SetUpProject-package}
\alias{SetUpProject}
\docType{package}
\title{
\packageTitle{SetUpProject}
}
\description{
\packageDescription{SetUpProject}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SetUpProject}
\packageIndices{SetUpProject}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{SetUpProject}

Maintainer: \packageMaintainer{SetUpProject}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
